rewrite build script and rework unsafe bindings
unsafe bindings are 1-to-1 with the raylib C API, name casing and all. the is to facilitate easier porting from C to Rust if one is willing to use unsafe.

because the entire surface of the unsafe API changed, the safe bindings have been removed to be rewritten because they were low quality as well.